Move create album into a new livewire table
